+++ include/env.hpp
@@ -14,10 +14,6 @@ std::string getEnv( std::string const & key , bool & exist)
     return std::string(val);
 }
 
-
-// template<typename T,
-// typename = std::enable_if_t<std::is_integral<T>::value>,
-// typename = std::enable_if_t<std::is_floating_point<T>::value>>
 template<typename T>
 T toType(std::string &s) {
     T value{};
@@ -36,38 +32,9 @@ T toType(std::string &s) {
     return value;
 }
 
-// std::string toType(std::string &s) {
-//     return s;
-// }
-
 struct serializable : refl::attr::usage::field
 {
 };
-
-struct Point
-{
-    std::string x;
-    std::string y;
-};
-
-REFL_TYPE(Point, bases<>)
-    REFL_FIELD(x, serializable()) // here we use serializable only as a maker
-    REFL_FIELD(y, serializable())
-REFL_END
-
-struct MyStruct
-{
-    std::string PATH;
-    std::string HOME;
-    int Age;
-};
-
-REFL_TYPE(MyStruct, bases<>)
-    REFL_FIELD(PATH, serializable())
-    REFL_FIELD(HOME, serializable())
-    REFL_FIELD(Age, serializable())
-REFL_END
-
 
 template <typename T>
 void Parse(T&& value)
@@ -99,22 +66,6 @@ void Parse(T&& value)
             }
             
             //member(value) = toType(val)
-
         }
     });
-}
-
-int main()
-{
-    // bool exists;
-    // std::string env = getEnv("PATH", exists);
-    // if(exists)
-    //     std::cout << "Your PATH is: " << env << '\n';
-    
-    MyStruct ms;
-    Parse(ms);
-
-    std::cout << "Your PATH is: " << ms.PATH << '\n';
-    std::cout << "Your HOME is: " << ms.HOME << '\n';
-    std::cout << "Your Age is: " << ms.Age << '\n';
 }
